backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
pillow
opencv-python-headless
numpy
torch
facenet-pytorch
sqlalchemy
aiofiles